% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultest.pe.chisq.R
\name{simultest.pe.chisq}
\alias{simultest.pe.chisq}
\title{Two-sample PE simultaneous test using chi-squared approximation}
\usage{
simultest.pe.chisq(dataX,dataY,delta_mean=NULL,delta_cov=NULL)
}
\arguments{
\item{dataX}{an \eqn{n_1} by \eqn{p} data matrix}

\item{dataY}{an \eqn{n_2} by \eqn{p} data matrix}

\item{delta_mean}{a scalar; the thresholding value used in the construction of
the PE component for mean test; see \code{\link{meantest.pe.comp}} for details.}

\item{delta_cov}{a scalar; the thresholding value used in the construction of
the PE component for covariance test; see \code{\link{covtest.pe.comp}} for details.}
}
\value{
\code{stat} the value of test statistic

\code{pval} the p-value for the test.
}
\description{
This function implements the two-sample PE simultaneous test on
high-dimensional mean vectors and covariance matrices using chi-squared approximation.
Suppose \eqn{\{\mathbf{X}_1, \ldots, \mathbf{X}_{n_1}\}} are i.i.d.
copies of \eqn{\mathbf{X}}, and \eqn{\{\mathbf{Y}_1, \ldots, \mathbf{Y}_{n_2}\}}
are i.i.d. copies of \eqn{\mathbf{Y}}.
Let \eqn{M_{PE}} and \eqn{T_{PE}} denote
the PE mean test statistic and PE covariance test statistic, respectively.
(see \code{\link{meantest.pe.comp}}
and \code{\link{covtest.pe.comp}} for details).
The PE simultaneous test statistic via chi-squared approximation is defined as
\deqn{S_{PE} = M_{PE}^2 + T_{PE}^2.}
It has been proved that with some regularity conditions, under the null hypothesis
\eqn{H_0: \boldsymbol{\mu}_1 = \boldsymbol{\mu}_2 \ \text{ and }
\ \mathbf{\Sigma}_1 = \mathbf{\Sigma}_2},
the two tests are asymptotically independent as \eqn{n_1, n_2, p\rightarrow \infty},
and therefore \eqn{S_{PE}} asymptotically converges in distribution to
a \eqn{\chi_2^2} distribution.
The asymptotic \eqn{p}-value is obtained by
\deqn{p\text{-value} = 1-F_{\chi_2^2}(S_{PE}),}
where \eqn{F_{\chi_2^2}(\cdot)} is the cdf of the \eqn{\chi_2^2} distribution.
}
\examples{
n1 = 100; n2 = 100; pp = 500
set.seed(1)
X = matrix(rnorm(n1*pp), nrow=n1, ncol=pp)
Y = matrix(rnorm(n2*pp), nrow=n2, ncol=pp)
simultest.pe.chisq(X,Y)
}
\references{
Yu, X., Li, D., Xue, L., and Li, R. (2022). Power-enhanced simultaneous test
of high-dimensional mean vectors and covariance matrices with application
to gene-set testing. \emph{Journal of the American Statistical Association},
(in press):1–14.
}
